+++ activerecord/lib/active_record/connection_adapters/postgresql_adapter.rb
@@ -370,6 +370,10 @@ module ActiveRecord
         true
       end
 
+      def supports_transaction_isolation?
+        true
+      end
+
       class StatementPool < ConnectionAdapters::StatementPool
         def initialize(connection, max)
           super